another one: daysValid switch to Long

+++ dataland-e2etests/src/test/kotlin/org/dataland/e2etests/utils/DatesHandler.kt
@@ -16,7 +16,7 @@ class DatesHandler {
         return SimpleDateFormat("MM/dd/yyyy").format(date)
     }
 
-    fun calculateExpectedExpiryDateSimpleFormatted(daysValid: Int? = null): String? {
+    fun calculateExpectedExpiryDateSimpleFormatted(daysValid: Long? = null): String? {
         return when (daysValid) {
             null -> null
             else -> formatDateAsSimpleDateString(addDaysToDate(Date(), daysValid))
